Added Missing Trips
Check "IMPORTANT.txt" for some important information on using this feed
in conjunction with the Massachusetts Bay Transportation Authority's
feed.

Also check trip_agency_id in trips.txt, which is outside the
specification. This field refers to the agency_id in agency.txt, but
indicates when a specific trip or trips within a route is operated by a
separate agency.